vm_server/requirements.txt
aiohttp==3.9.1
paho-mqtt==1.6.1
python-dotenv==1.0.0
asyncio-mqtt==0.16.1
